requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
coreapi==2.3.3
django-filter==24.2
setuptools==70.1.1
black==24.4.2
python-dateutil==2.9.0